test(utils): tighten cloneDeep warn-once test and pin null-merge preconditions
Address Copilot review on PR #111:

- form-utils.spec.ts null-target regression test: explicitly assert
  the Angular preconditions (raw value reports the underlying value
  for the disabled controls) so the test fails loudly if Angular's
  disabled-control behavior shifts.
- form-utils.cloneDeep-deprecation.spec.ts: extracted the warn-once
  test into its own isolated spec file. Vitest gives each file a
  fresh module instance, so the module-level flag starts unset and
  we can assert exactly one warning on first call plus zero further
  warnings across many subsequent calls (instead of the previous
  weak `<= 1` assertion).
- form-utils.ts cloneDeep JSDoc: redirect the removal-tracking link
  from #107 (which is the fastDeepEqual cycle-detection issue) to
  PRD-bug-sweep.md Bundle D, where cloneDeep removal is actually
  documented.
